Add sourcemap support
Unfortunately when running tests in `watch` mode, I see "...WARN [reporter]: SourceMap position not
found for trace: undefined". I think this may be due to remote dependencies not having a sourcemap.
We could just ignore this warning, but it's annoying. Perhaps we could use `formatError` from
http://karma-runner.github.io/1.0/config/configuration-file.html to filter out the warning, but then
we run the risk of filtering out warnings caused by the user of erik not including sourcemaps.